+++ src/main/java/io/sentry/SentryInstallerLifecycleParticipant.java
@@ -22,6 +22,7 @@ import io.sentry.autoinstall.graphql.GraphqlInstallStrategy;
 import io.sentry.autoinstall.jdbc.JdbcInstallStrategy;
 import io.sentry.autoinstall.log4j2.Log4j2InstallStrategy;
 import io.sentry.autoinstall.logback.LogbackInstallStrategy;
+import io.sentry.autoinstall.opentelemetry.OpenTelemetryBomInstaller;
 import io.sentry.autoinstall.quartz.QuartzInstallStrategy;
 import io.sentry.autoinstall.spring.*;
 import io.sentry.autoinstall.util.ManagedSentryVersionResolver;
@@ -73,6 +74,14 @@ public class SentryInstallerLifecycleParticipant extends AbstractMavenLifecycleP
   @SuppressWarnings("NullAway")
   @Inject
   private @NotNull BuildPluginManager pluginManager;
+
+  @SuppressWarnings("NullAway")
+  @Inject
+  private @NotNull org.apache.maven.project.ProjectBuilder projectBuilder;
+
+  @SuppressWarnings("NullAway")
+  @Inject
+  private @NotNull org.apache.maven.repository.RepositorySystem mavenRepositorySystem;
 
   private static final @NotNull org.slf4j.Logger logger =
       LoggerFactory.getLogger(SentryInstallerLifecycleParticipant.class);
@@ -137,6 +146,22 @@ public class SentryInstallerLifecycleParticipant extends AbstractMavenLifecycleP
             installer.install(dependencyList, resolvedArtifacts, autoInstallState);
           } catch (Throwable e) {
             logger.error("Unable to instantiate installer class: " + installerClass.getName(), e);
+          }
+        }
+
+        if (!pluginConfig.isSkipInstallOpenTelemetryBom()) {
+          try {
+            new OpenTelemetryBomInstaller()
+                .install(
+                    project,
+                    resolvedArtifacts,
+                    sentryVersion,
+                    projectBuilder,
+                    mavenRepositorySystem,
+                    session.getProjectBuildingRequest());
+          } catch (Throwable e) {
+            logger.error(
+                "Unable to install " + OpenTelemetryBomInstaller.SENTRY_OPENTELEMETRY_BOM_ID, e);
           }
         }
       } catch (Throwable t) {